refactor: Fix Fonts and Implement Better Responsivity
