+++ src/Home.tsx
@@ -1,13 +1,14 @@
 import { ChakraProvider } from '@chakra-ui/react'
-import "@assets/fonts.css"
-import Header from '@components/Header/header';
-import Footer from '@components/Footer/footer';
+import Header from '@components/Header/index';
+import Footer from '@components/Footer/index';
+import Presentation from '@components/Presentation/index'
 
 function App() {
   
   return (
     <ChakraProvider>
         <Header/>
+        <Presentation/>
         <Footer/>
     </ChakraProvider>
   )
