+++ src/components/Footer/index.tsx
@@ -1,5 +1,5 @@
 import { Box, Text } from "@chakra-ui/react"
-import styles from "./footer.module.css"
+import styles from "./index.module.css"
 
 export default function Index() {
